% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{find_pkgs}
\alias{find_pkgs}
\title{Determine which package a function is from}
\usage{
find_pkgs(funs)
}
\arguments{
\item{funs}{a character vector of function names}
}
\value{
dataframe with the function (fun) and it's parent package (package),
each as strings
}
\description{
This function takes a character vector of function names and returns a dataframe
containing the package containing those functions.
}
\details{
Objects that appear to be in the global environment will be indicated as `.GlobalEnv`.
Functions that are referenced via a namespace (i.e. `::` or `:::`)
may not indicate a package, particularly when the package is not loaded.

Where the same function is found in multiple packages, all packages are returned.
This is particularly the case where functions are re-exported by multiple packages
(e.g. `dplyr::mutate` is also exported by `gtsummary`).

Where the `conflicted` package is used, the package that takes priority is returned.
}
\examples{
find_pkgs("mutate")
}
